scala 2.13 support
the rummage/try code fails to cross-build to 2.13 due to the collections overhaul removing CanBuildFrom and IterableLike, which is explictly used, and would be moderately difficult to duplicate.

but this code is completely unused within the library so just remove it along with its test.

the -Ywarn-nullary-unit option is removed because its behavior is covered by -Xlint but is an invalid option in 2.13.